+++ src/Sulu/Bundle/ContactBundle/Admin/ContactAdmin.php
@@ -62,14 +62,6 @@ class ContactAdmin extends Admin
     /**
      * {@inheritdoc}
      */
-    public function getCommands()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getJsBundleName()
     {
         return 'sulucontact';