+++ src/Sulu/Bundle/ContentBundle/Admin/ContentAdmin.php
@@ -98,14 +98,6 @@ class ContentAdmin extends Admin
     /**
      * {@inheritdoc}
      */
-    public function getCommands()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getJsBundleName()
     {
         return 'sulucontent';