+++ src/Sulu/Bundle/MediaBundle/Admin/MediaAdmin.php
@@ -62,14 +62,6 @@ class MediaAdmin extends Admin
     /**
      * {@inheritdoc}
      */
-    public function getCommands()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getJsBundleName()
     {
         return 'sulumedia';