+++ src/Sulu/Bundle/SearchBundle/Admin/SuluSearchAdmin.php
@@ -39,14 +39,6 @@ class SuluSearchAdmin extends Admin
     /**
      * {@inheritdoc}
      */
-    public function getCommands()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getJsBundleName()
     {
         return 'sulusearch';