+++ src/Sulu/Bundle/WebsiteBundle/Admin/WebsiteAdmin.php
@@ -40,14 +40,6 @@ class WebsiteAdmin extends Admin
     /**
      * {@inheritdoc}
      */
-    public function getCommands()
-    {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getJsBundleName()
     {
         return 'suluwebsite';